+++ src/completion-renderers-zsh.lisp
@@ -36,9 +36,15 @@
       (format out "      ;;~%"))))
 
 (defun render-zsh-completion (app &optional stream)
-  "Render a zsh completion script."
-  (let ((stream (or stream *standard-output*))
-        (function-name (%completion-function-name app))
+  "Render a zsh completion script.
+
+With no STREAM, return the completion script as a string. With a STREAM,
+write the script to it and return no values."
+  (unless stream
+    (return-from render-zsh-completion
+      (with-output-to-string (string-stream)
+        (render-zsh-completion app string-stream))))
+  (let ((function-name (%completion-function-name app))
         (app-name (app-name app)))
     (format stream "#compdef ~A~%" app-name)
     (format stream "~A() {~%" function-name)